Changes to be committed: 	modified:   Sites/collex/app/assets/javascripts/full_record.js 	modified:   Sites/collex/app/assets/javascripts/jquery.viafauto.extended.js
I fixed full_record.js so that all of the facet information now shows up at the top of the page. This happens in the createResultContents() function.  This was done by copying the complete function to a new one and modifying there.  The depricated function is still there called createResultContentsOrigninal().  I also looked into an error that the viaf autocomplete is throwing, but I didn't change anything.