Changes to be committed: 	modified:   ../Sites/catalog/app/models/query_format.rb 	modified:   ../Sites/catalog/app/models/solr.rb 	modified:   ../Sites/collex/app/assets/javascripts/search/search_result_draw_query.js 	modified:   ../Sites/collex/app/controllers/search_controller.rb 	modified:   ../Sites/collex/app/models/catalog.rb
Committing to preserve current set of changes.
Half way through fixing the problem of searching for ESTC
numbers.  Been working in search_cotnroller.rb (collex)
in the process_constraints(query) function.  Added code
that looks to see if there is a URI facet filter submitted
and, if there is, it removes the q= facet and replaces
it with one for the estc ID and then adds the constraint.
This approach fails the q= checking parameter in catalog
query_format model.  I do see that there is a URI format
option in query_format, so I'm going to try just
converting the q= q=*.* and then passing on the URI
parameter to the search.